# apps/backend/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Plataforma SaaS especializada em cafeicultura para gestão operacional, produtiva, financeira e analítica de propriedades cafeeiras.",
        "title": "CafeOS API",
        "contact": {},
        "license": {
            "name": "Proprietary"
        },
        "version": "0.1.0"
    },
    "host": "localhost:5001",
    "basePath": "/api/v1/{tenant_id}",
    "paths": {
...
        "/api/v1/{tenant_id}/farms/{farm_id}/plots": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lista todos os talhões pertencentes a uma fazenda",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "plots (Talhões)"
                ],
                "summary": "Listar talhões por fazenda",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Tenant",
                        "name": "tenant_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID da Fazenda",
                        "name": "farm_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.SwaggerPlot"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/{tenant_id}/plots": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lista todos os talhões de todas as fazendas do tenant",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "plots (Talhões)"
                ],
                "summary": "Listar todos os talhões",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Tenant",
                        "name": "tenant_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.SwaggerPlot"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Registra um novo talhão em uma fazenda",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "plots (Talhões)"
                ],
                "summary": "Criar talhão",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Tenant",
                        "name": "tenant_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados do talhão",
                        "name": "plot",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.createPlotRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.SwaggerPlot"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/{tenant_id}/plots/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna um único talhão",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "plots (Talhões)"
                ],
                "summary": "Obter talhão por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Tenant",
                        "name": "tenant_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID do Talhão",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.SwaggerPlot"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Atualiza dados do talhão por ID (atualização parcial - somente os campos informados são alterados)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "plots (Talhões)"
                ],
                "summary": "Atualizar talhão",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Tenant",
                        "name": "tenant_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID do Talhão",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados atualizados do talhão",
                        "name": "plot",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.SwaggerPlot"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.SwaggerPlot"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Exclui um talhão por ID",
                "tags": [
                    "plots (Talhões)"
                ],
                "summary": "Excluir talhão",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Tenant",
                        "name": "tenant_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID do Talhão",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.PlotStage": {
            "type": "string",
            "enum": [
                "formacao",
                "producao"
            ],
            "x-enum-varnames": [
                "PlotStageFormacao",
                "PlotStageProducao"
            ]
        },
...
        "handler.SwaggerPlot": {
            "type": "object",
            "properties": {
                "activation_date": {
                    "type": "string"
                },
                "altitude": {
                    "type": "integer"
                },
                "app_area_ha": {
                    "type": "number"
                },
                "area_ha": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "crop_type": {
                    "type": "string"
                },
                "cultivar": {
                    "type": "string"
                },
                "dam_area_ha": {
                    "type": "number"
                },
                "deactivation_date": {
                    "type": "string"
                },
                "farm_id": {
                    "type": "string"
                },
                "formation_cost_per_ha": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "improvements_area_ha": {
                    "type": "number"
                },
                "intercropped": {
                    "type": "boolean"
                },
                "irrigation": {
                    "type": "string"
                },
                "leased": {
                    "type": "boolean"
                },
                "legal_reserve_area_ha": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "plant_count": {
                    "type": "integer"
                },
                "plant_spacing_m": {
                    "type": "number"
                },
                "planting_date": {
                    "type": "string"
                },
                "planting_year": {
                    "type": "integer"
                },
                "roads_area_ha": {
                    "type": "number"
                },
                "row_spacing_m": {
                    "type": "number"
                },
                "secondary_crop": {
                    "type": "string"
                },
                "soil_type": {
                    "type": "string"
                },
                "stage": {
                    "$ref": "#/definitions/entity.PlotStage"
                },
                "tenant_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "useful_life_years": {
                    "type": "integer"
                }
            }
        },
...
        "handler.createPlotRequest": {
            "type": "object",
            "properties": {
                "activation_date": {
                    "type": "string"
                },
                "altitude": {
                    "type": "integer"
                },
                "app_area_ha": {
                    "type": "number"
                },
                "area_ha": {
                    "type": "number"
                },
                "crop_type": {
                    "type": "string"
                },
                "cultivar": {
                    "type": "string"
                },
                "dam_area_ha": {
                    "type": "number"
                },
                "deactivation_date": {
                    "type": "string"
                },
                "farm_id": {
                    "type": "string"
                },
                "formation_cost_per_ha": {
                    "type": "number"
                },
                "improvements_area_ha": {
                    "type": "number"
                },
                "intercropped": {
                    "type": "boolean"
                },
                "irrigation": {
                    "type": "string"
                },
                "leased": {
                    "type": "boolean"
                },
                "legal_reserve_area_ha": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "plant_count": {
                    "type": "integer"
                },
                "plant_spacing_m": {
                    "type": "number"
                },
                "planting_date": {
                    "type": "string"
                },
                "planting_year": {
                    "type": "integer"
                },
                "roads_area_ha": {
                    "type": "number"
                },
                "row_spacing_m": {
                    "type": "number"
                },
                "secondary_crop": {
                    "type": "string"
                },
                "soil_type": {
                    "type": "string"
                },
                "stage": {
                    "type": "string"
                },
                "useful_life_years": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT token. Use: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}